models: overwrite `find`, `find_by` for custom urls
Makes `find`-helpers accept a `custom_url` in place of an `uuid`.

Only the common use cases are covered,
to guard against silent bug it throws on unsupported usage.

Reason: The way CustomUrls are integrated with Rails,
they are completely transparent to individual controller actions
(e.g. `params[id] == 'my-custom-url'`).

This change was deemed the least invasive and error-prone
(as opposed to having to handle it explictly in the router as well as each action).

- alias_method_chain find :-o
- #find_by_id
- disclaimers
- model specs
- relations, /w specs
- find_by and find_by! prevention

Signed-off-by: Max Albrecht <max.albrecht@zhdk.ch>